separate knowledgemap from coachreports.
